Move the anchor for bosh-deployment out of dev;
This patch moves the anchor for all bosh-deployments out of Development
and into Staging. This is temporary and we can revert this specific
commit when we get CredHub on all environments. For now, I don't see a
better way to make sure that we only deploy `credhub.yml` in a
particular environment without doing this.